chore(virtual-event): adjust component props types for React 18, fix StreamGeneric type
