+++ virtual-event/src/contexts/GiphyContext.tsx
@@ -1,10 +1,11 @@
 import React, { useContext, useState } from 'react';
+import { PropsWithChildrenOnly } from '../types';
 
 const GiphyContext = React.createContext(
   {} as { giphyState: boolean; setGiphyState: React.Dispatch<React.SetStateAction<boolean>> },
 );
 
-export const GiphyContextProvider: React.FC = ({ children }) => {
+export const GiphyContextProvider= ({ children }: PropsWithChildrenOnly) => {
   const [giphyState, setGiphyState] = useState(false);
 
   const value = { giphyState, setGiphyState };
