+++ virtual-event/src/types.ts
@@ -1,0 +1,21 @@
+import { PropsWithChildren } from 'react';
+import { LiteralStringForUnion, UR } from 'stream-chat';
+
+export type PropsWithChildrenOnly = PropsWithChildren<Record<never, never>>;
+type AttachmentType = UR;
+type ChannelType = UR;
+type CommandType = LiteralStringForUnion;
+type EventType = UR;
+type MessageType = UR & { up_votes?: string[] };
+type ReactionType = UR;
+export type UserType = UR & { image?: string; title?: string };
+
+export type StreamChatType = {
+  attachmentType: AttachmentType;
+  channelType: ChannelType;
+  commandType: CommandType;
+  eventType: EventType;
+  messageType: MessageType;
+  reactionType: ReactionType;
+  userType: UserType;
+}
